# openapi.yml
openapi: 3.0.3
info:
  title: Reminder API
  description: Apple Reminder 스타일 리마인더 관리 API
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /api/reminders:
    get:
      summary: 리마인더 조회
      description: 전체 조회, 필터 조회, 검색을 지원한다. search와 filter가 동시에 주어지면 search가 우선한다.
      operationId: getReminders
      parameters:
        - name: filter
          in: query
          required: false
          description: "필터 종류: today, scheduled, completed"
          schema:
            type: string
            enum: [today, scheduled, completed]
        - name: search
          in: query
          required: false
          description: 제목/메모 검색 키워드
          schema:
            type: string
      responses:
        "200":
          description: 리마인더 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReminderResponse"

    post:
      summary: 리마인더 생성
      operationId: createReminder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderRequest"
      responses:
        "201":
          description: 생성된 리마인더
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReminderResponse"

  /api/reminders/counts:
    get:
      summary: 필터별 카운트 조회
      operationId: getCounts
      responses:
        "200":
          description: 필터별 리마인더 수
          content:
            application/json:
              schema:
                type: object
                properties:
                  today:
                    type: integer
                    format: int64
                  scheduled:
                    type: integer
                    format: int64
                  all:
                    type: integer
                    format: int64
                  completed:
                    type: integer
                    format: int64

  /api/reminders/{id}:
    get:
      summary: 리마인더 단건 조회
      operationId: getReminder
      parameters:
        - $ref: "#/components/parameters/ReminderId"
      responses:
        "200":
          description: 리마인더 상세
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReminderResponse"
        "404":
          description: 리마인더를 찾을 수 없음

    put:
      summary: 리마인더 수정
      operationId: updateReminder
      parameters:
        - $ref: "#/components/parameters/ReminderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderRequest"
      responses:
        "200":
          description: 수정된 리마인더
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReminderResponse"
        "404":
          description: 리마인더를 찾을 수 없음

    delete:
      summary: 리마인더 삭제
      operationId: deleteReminder
      parameters:
        - $ref: "#/components/parameters/ReminderId"
      responses:
        "204":
          description: 삭제 완료
        "404":
          description: 리마인더를 찾을 수 없음

  /api/reminders/{id}/complete:
    patch:
      summary: 리마인더 완료/미완료 토글
      operationId: toggleComplete
      parameters:
        - $ref: "#/components/parameters/ReminderId"
      responses:
        "200":
          description: 토글된 리마인더
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReminderResponse"
        "404":
          description: 리마인더를 찾을 수 없음

components:
  parameters:
    ReminderId:
      name: id
      in: path
      required: true
      description: 리마인더 ID
      schema:
        type: integer
        format: int64

  schemas:
    ReminderRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          description: 리마인더 제목 (필수)
          example: 장보기
        description:
          type: string
          nullable: true
          description: 리마인더 메모
          example: 우유, 빵 사기
        remindAt:
          type: string
          format: date-time
          nullable: true
          description: 알림 일시
          example: "2026-03-20T14:00:00"

    ReminderResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: 장보기
        description:
          type: string
          nullable: true
          example: 우유, 빵 사기
        remindAt:
          type: string
          format: date-time
          nullable: true
          example: "2026-03-20T14:00:00"
        completed:
          type: boolean
          example: false
        createdAt:
          type: string
          format: date-time
          example: "2026-03-16T10:30:00"