# src/main/java/logan/ai/reminder/controller/ReminderResponse.java
package logan.ai.reminder.controller;

import logan.ai.reminder.entity.Reminder;

import java.time.LocalDateTime;

public record ReminderResponse(
        Long id,
        String title,
        String description,
        LocalDateTime remindAt,
        Boolean completed,
        LocalDateTime createdAt
) {
    public static ReminderResponse from(Reminder reminder) {
        return new ReminderResponse(
                reminder.getId(),
                reminder.getTitle(),
                reminder.getDescription(),
                reminder.getRemindAt(),
                reminder.getCompleted(),
                reminder.getCreatedAt()
        );
    }
}